created reusable GreenModal component
