+++ src/components/GreenModal/icons/CloseIcon.tsx
