+++ src/containers/Airdrop/AirdropGiftTracker/components/Claimmodal/ClaimModal.tsx
@@ -1,9 +1,6 @@
 import {
     ActionWrapper,
-    BoxWrapper,
     ClaimButton,
-    CloseButton,
-    Cover,
     FinePrint,
     Gem1,
     Gem2,
@@ -16,7 +13,6 @@ import {
     Text,
     TextWrapper,
     Title,
-    Wrapper,
     XLogo,
 } from './styles';
 import gemImage from './images/gems.png';
@@ -27,7 +23,7 @@ import { Trans, useTranslation } from 'react-i18next';
 import { GemImage } from '../Gems/styles';
 import XLogoIcon from './icons/XLogoIcon';
 import { useAppConfigStore } from '@app/store/useAppConfigStore';
-import CloseIcon from './icons/CloseIcon';
+import GreenModal from '@app/components/GreenModal/GreenModal';
 
 interface ClaimModalProps {
     onSubmit: (code?: string) => void;
@@ -48,61 +44,53 @@ export default function ClaimModal({ onSubmit, onClose }: ClaimModalProps) {
     }, [claimCode, onSubmit, setAllowTelemetry]);
 
     return (
-        <Wrapper>
-            <BoxWrapper initial={{ opacity: 0, y: '100px' }} animate={{ opacity: 1, y: 0 }} exit={{ opacity: 0 }}>
-                <CloseButton onClick={onClose}>
-                    <CloseIcon />
-                </CloseButton>
+        <GreenModal onClose={onClose}>
+            <GemsWrapper>
+                <Gem1 src={gemLargeImage} alt="" />
+                <Gem2 src={gemLargeImage} alt="" />
+                <Gem3 src={gemLargeImage} alt="" />
+            </GemsWrapper>
 
-                <GemsWrapper>
-                    <Gem1 src={gemLargeImage} alt="" />
-                    <Gem2 src={gemLargeImage} alt="" />
-                    <Gem3 src={gemLargeImage} alt="" />
-                </GemsWrapper>
+            <TextWrapper>
+                <Title>{t('claimModalTitle')}</Title>
 
-                <TextWrapper>
-                    <Title>{t('claimModalTitle')}</Title>
+                <Text
+                    dangerouslySetInnerHTML={{
+                        __html: t('claimModalText', {
+                            gems1: (referralQuestPoints?.pointsForClaimingReferral || GIFT_GEMS).toLocaleString(),
+                            gems2: MAX_GEMS.toLocaleString(),
+                        }),
+                    }}
+                />
+            </TextWrapper>
 
-                    <Text
-                        dangerouslySetInnerHTML={{
-                            __html: t('claimModalText', {
-                                gems1: (referralQuestPoints?.pointsForClaimingReferral || GIFT_GEMS).toLocaleString(),
-                                gems2: MAX_GEMS.toLocaleString(),
-                            }),
-                        }}
+            <ActionWrapper>
+                <InputWrapper>
+                    <InputLabel>{t('claimModalFieldLabel')}</InputLabel>
+                    <StyledInput
+                        type="text"
+                        placeholder={t('claimModalFieldPlaceholder')}
+                        onChange={(e) => setClaimCode(e.target.value)}
+                        value={claimCode}
                     />
-                </TextWrapper>
+                    <InputGems>
+                        {GIFT_GEMS.toLocaleString()} <GemImage src={gemImage} alt="" />
+                    </InputGems>
+                </InputWrapper>
 
-                <ActionWrapper>
-                    <InputWrapper>
-                        <InputLabel>{t('claimModalFieldLabel')}</InputLabel>
-                        <StyledInput
-                            type="text"
-                            placeholder={t('claimModalFieldPlaceholder')}
-                            onChange={(e) => setClaimCode(e.target.value)}
-                            value={claimCode}
-                        />
-                        <InputGems>
-                            {GIFT_GEMS.toLocaleString()} <GemImage src={gemImage} alt="" />
-                        </InputGems>
-                    </InputWrapper>
+                <ClaimButton onClick={handleSubmit}>
+                    {t('claimGems')}
+                    <XLogo>
+                        <XLogoIcon />
+                    </XLogo>
+                </ClaimButton>
 
-                    <ClaimButton onClick={handleSubmit}>
-                        {t('claimGems')}
-                        <XLogo>
-                            <XLogoIcon />
-                        </XLogo>
-                    </ClaimButton>
-
-                    {!allowTelemetry && (
-                        <FinePrint>
-                            <Trans t={t} i18nKey="claimModalFinePrint" ns="airdrop" components={{ bold: <strong /> }} />
-                        </FinePrint>
-                    )}
-                </ActionWrapper>
-            </BoxWrapper>
-
-            <Cover onClick={onClose} initial={{ opacity: 0 }} animate={{ opacity: 1 }} exit={{ opacity: 0 }} />
-        </Wrapper>
+                {!allowTelemetry && (
+                    <FinePrint>
+                        <Trans t={t} i18nKey="claimModalFinePrint" ns="airdrop" components={{ bold: <strong /> }} />
+                    </FinePrint>
+                )}
+            </ActionWrapper>
+        </GreenModal>
     );
 }
